chore: decompose sqllog files
