+++ pkg/sqllog/sqllog.go
@@ -47,23 +47,6 @@ func init() {
 	}
 }
 
-type Driver interface {
-	List(ctx context.Context, key, rangeEnd []byte, limit, revision int64) (*sql.Rows, error)
-	ListTTL(ctx context.Context, revision int64) (*sql.Rows, error)
-	Count(ctx context.Context, key, rangeEnd []byte, revision int64) (int64, error)
-	CurrentRevision(ctx context.Context) (int64, error)
-	AfterPrefix(ctx context.Context, key, rangeEnd []byte, fromRevision, toRevision int64) (*sql.Rows, error)
-	After(ctx context.Context, startRevision, limit int64) (*sql.Rows, error)
-	Create(ctx context.Context, key []byte, value []byte, lease int64) (int64, bool, error)
-	Update(ctx context.Context, key []byte, value []byte, prevRev, lease int64) (int64, bool, error)
-	Delete(ctx context.Context, key []byte, revision int64) (int64, bool, error)
-	DeleteRevision(ctx context.Context, revision int64) error
-	GetCompactRevision(ctx context.Context) (int64, int64, error)
-	Compact(ctx context.Context, revision int64) error
-	GetSize(ctx context.Context) (int64, error)
-	Close() error
-}
-
 type SQLLog struct {
 	mu sync.Mutex
 
@@ -632,79 +615,6 @@ func scanEvent(rows *sql.Rows) (*limited.Event, error) {
 	return event, nil
 }
 
-type watcherGroup struct {
-	mu              sync.Mutex
-	ctx             context.Context
-	driver          Driver
-	currentRevision int64
-	watchers        map[int64]*watcher
-	updates         chan limited.WatcherGroupUpdate
-	stop            func() bool
-}
-
-type watcher struct {
-	watchId       int64
-	key, rangeEnd []byte
-	initialized   bool
-	events        []*limited.Event
-}
-
-type watcherGroupUpdate struct {
-	revision int64
-	updates  []limited.WatcherUpdate
-}
-
-func (w *watcherGroupUpdate) Revision() int64                   { return w.revision }
-func (w *watcherGroupUpdate) Watchers() []limited.WatcherUpdate { return w.updates }
-
-func (w *watcherGroup) publish(currentRevision int64, events []*limited.Event) bool {
-	_, span := otelTracer.Start(w.ctx, fmt.Sprintf("%s.publish", otelName))
-	span.SetAttributes(attribute.Int64("currentRevision", currentRevision))
-	span.SetAttributes(attribute.Int("events", len(events)))
-	span.SetAttributes(attribute.Int("watchers", len(w.watchers)))
-	span.SetAttributes(attribute.Int("updates", len(w.updates)))
-	defer span.End()
-	w.mu.Lock()
-	defer w.mu.Unlock()
-
-	w.currentRevision = currentRevision
-	updates := []limited.WatcherUpdate{}
-	for watchId, watcher := range w.watchers {
-		events := watcher.update(events)
-		if len(events) > 0 {
-			updates = append(updates, limited.WatcherUpdate{
-				WatcherId: watchId,
-				Events:    events,
-			})
-		}
-	}
-
-	groupUpdate := &watcherGroupUpdate{
-		revision: currentRevision,
-		updates:  updates,
-	}
-	select {
-	case w.updates <- groupUpdate:
-		return true
-	default:
-		return false
-	}
-}
-
-func (w *watcher) update(events []*limited.Event) []*limited.Event {
-	if !w.initialized {
-		w.events = appendEvents(events, w.events, w.key, w.rangeEnd)
-		return nil
-	} else if len(w.events) > 0 {
-		initialEvents := w.events
-		w.events = nil
-		return appendEvents(events, initialEvents, w.key, w.rangeEnd)
-	}
-
-	filtered := make([]*limited.Event, 0, len(events))
-	return appendEvents(events, filtered, w.key, w.rangeEnd)
-}
-
 func appendEvents(src, dst []*limited.Event, key, rangeEnd []byte) []*limited.Event {
 	for _, event := range src {
 		if bytes.Compare(key, event.Kv.Key) <= 0 && bytes.Compare(rangeEnd, event.Kv.Key) > 0 {
@@ -713,94 +623,3 @@ func appendEvents(src, dst []*limited.Event, key, rangeEnd []byte) []*limited.Ev
 	}
 	return dst
 }
-
-func (w *watcherGroup) Watch(watchId int64, key, rangeEnd []byte, startRevision int64) error {
-	var err error
-	ctx, span := otelTracer.Start(w.ctx, fmt.Sprintf("%s.Watch", otelName))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-
-	span.SetAttributes(
-		attribute.String("key", string(key)),
-		attribute.String("rangeEnd", string(rangeEnd)),
-		attribute.Int64("startRevision", startRevision),
-		attribute.Int64("watchId", watchId),
-	)
-
-	watcher := &watcher{
-		watchId:  watchId,
-		key:      key,
-		rangeEnd: rangeEnd,
-	}
-
-	w.mu.Lock()
-	w.watchers[watchId] = watcher
-	if startRevision > 0 {
-		currentRevision := w.currentRevision
-		w.mu.Unlock()
-
-		compactRev, rev, err := w.driver.GetCompactRevision(ctx)
-		if err != nil {
-			logrus.Errorf("Failed to get compact revision: %v", err)
-			return err
-		}
-		if startRevision < compactRev {
-			return &limited.CompactedError{CompactRevision: compactRev, CurrentRevision: rev}
-		}
-
-		initialEvents, err := w.initialEvents(ctx, key, rangeEnd, startRevision-1, currentRevision)
-		if err != nil {
-			w.mu.Lock()
-			delete(w.watchers, watchId)
-			w.mu.Unlock()
-			if !errors.Is(err, context.Canceled) {
-				logrus.Errorf("Failed to list %s for revision %d: %v", key, startRevision, err)
-			}
-			return err
-		}
-
-		w.mu.Lock()
-		watcher.events = append(initialEvents, watcher.events...)
-	}
-	watcher.initialized = true
-	w.mu.Unlock()
-
-	return nil
-}
-
-func (w *watcherGroup) initialEvents(ctx context.Context, key, rangeEnd []byte, fromRevision, toRevision int64) ([]*limited.Event, error) {
-	var err error
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.initialEvents", otelName))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-	span.SetAttributes(
-		attribute.String("key", string(key)),
-		attribute.String("rangeEnd", string(rangeEnd)),
-		attribute.Int64("fromRevision", fromRevision),
-		attribute.Int64("toRevision", toRevision),
-	)
-
-	rows, err := w.driver.AfterPrefix(ctx, key, rangeEnd, fromRevision, toRevision)
-	if err != nil {
-		return nil, err
-	}
-
-	events, err := ScanAll(rows, scanEvent)
-	span.SetAttributes(attribute.Int("events", len(events)))
-	return events, err
-}
-
-func (w *watcherGroup) Unwatch(watchId int64) {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-
-	delete(w.watchers, watchId)
-}
-
-func (w *watcherGroup) Updates() <-chan limited.WatcherGroupUpdate { return w.updates }
-
-var _ limited.WatcherGroup = &watcherGroup{}
